Crudely disect command
