+++ code/src/healthstatus/__main__.py
@@ -1,475 +1,25 @@
-#!/usr/bin/env python3
 from __future__ import annotations
-from collections import defaultdict, deque
-from collections.abc import AsyncIterator, Awaitable, Callable, Sequence
-from dataclasses import dataclass, field
-from datetime import datetime
-from enum import Enum
-import io
 import logging
-import math
 import os
-import pathlib
-import re
 from shutil import rmtree
 from signal import SIGINT
 import subprocess
-import sys
 import tempfile
-import textwrap
 from time import sleep
-from typing import Optional
 import anyio
-from anyio.streams.memory import MemoryObjectReceiveStream
 import click
 import requests
-from ruamel.yaml import YAML
+from .checker import HealthStatus
+from .config import MATNWB_INSTALL_DIR
+from .core import log
 
-if sys.version_info[:2] >= (3, 8):
-    from importlib.metadata import version
-else:
-    from importlib_metadata import version
 
-MATNWB_INSTALL_DIR = pathlib.Path(__file__).with_name("matnwb")
+@click.group()
+def main() -> None:
+    pass
 
-MATNWB_SAVEDIR = "/mnt/fast/dandi/dandisets-healthstatus"
 
-PACKAGES_TO_VERSION = ["pynwb", "hdmf"]
-
-PYNWB_OPEN_LOAD_NS_SCRIPT = anyio.Path(__file__).with_name("pynwb_open_load_ns.py")
-
-TIMEOUT = 3600
-
-WORKERS_PER_DANDISET = 1
-
-log = logging.getLogger()
-
-
-async def run_test_command(
-    testname: str,
-    asset: Asset,
-    command: list[str],
-    env: Optional[dict[str, str]] = None,
-) -> TestResult:
-    if env is not None:
-        env = {**os.environ, **env}
-    try:
-        async with anyio.fail_after(TIMEOUT):
-            r = await anyio.run_process(
-                command,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT,
-                check=False,
-                env=env,
-            )
-    except TimeoutError:
-        return TestResult(testname=testname, asset=asset, outcome=Outcome.TIMEOUT)
-    else:
-        if r.returncode == 0:
-            return TestResult(testname=testname, asset=asset, outcome=Outcome.PASS)
-        else:
-            return TestResult(
-                testname=testname,
-                asset=asset,
-                outcome=Outcome.FAIL,
-                output=r.stdout.decode("utf-8", "surrogateescape"),
-            )
-
-
-async def pynwb_open_load_ns(asset: Asset) -> TestResult:
-    return await run_test_command(
-        "pynwb_open_load_ns",
-        asset,
-        [sys.executable, str(PYNWB_OPEN_LOAD_NS_SCRIPT), str(asset.filepath)],
-    )
-
-
-async def matnwb_nwbRead(asset: Asset) -> TestResult:
-    return await run_test_command(
-        "matnwb_nwbRead",
-        asset,
-        [
-            "matlab",
-            "-nodesktop",
-            "-batch",
-            (
-                f"nwb = nwbRead({str(asset.filepath)!r}, 'savedir',"
-                f" {MATNWB_SAVEDIR!r})"
-            ),
-        ],
-        env={"MATLABPATH": str(MATNWB_INSTALL_DIR)},
-    )
-
-
-TESTS = [pynwb_open_load_ns, matnwb_nwbRead]
-TEST_NAMES = [t.__name__ for t in TESTS]
-
-
-@dataclass
-class TestCase:
-    asset: Asset
-    testfunc: Callable[[Asset], Awaitable[TestResult]]
-
-    async def run(self) -> TestResult:
-        r = await self.testfunc(self.asset)
-        log.info(
-            "Dandiset %s, asset %s, test %s: %s",
-            self.asset.dandiset_id,
-            self.asset.asset_path,
-            r.testname,
-            r.outcome.name,
-        )
-        return r
-
-
-class Outcome(Enum):
-    PASS = "pass"
-    FAIL = "fail"
-    TIMEOUT = "timeout"
-
-
-@dataclass
-class TestResult:
-    testname: str
-    asset: Asset
-    outcome: Outcome
-    output: Optional[str] = None
-
-
-@dataclass
-class RunData:
-    timestamp: datetime
-    versions: dict[str, str]
-
-    @classmethod
-    def get(cls) -> RunData:
-        return cls(
-            timestamp=datetime.now().astimezone(),
-            versions=get_package_versions(),
-        )
-
-
-@dataclass
-class HealthStatus:
-    backup_root: anyio.Path
-    reports_root: anyio.Path
-    run_data: RunData = field(default_factory=RunData.get)
-
-    async def run(self, dandisets: Sequence[str], dandiset_jobs: int) -> None:
-        all_reports: list[DandisetReport] = []
-
-        async def dowork(rec: MemoryObjectReceiveStream[Dandiset]) -> None:
-            async with rec:
-                async for dandiset in rec:
-                    log.info("Processing Dandiset %s", dandiset.identifier)
-                    report = await dandiset.test_assets()
-                    await report.dump(
-                        self.reports_root / "results" / dandiset.identifier,
-                        self.run_data,
-                    )
-                    all_reports.append(report)
-
-        async with anyio.create_task_group() as tg:
-            sender, receiver = anyio.create_memory_object_stream(math.inf)
-            async with receiver:
-                for _ in range(dandiset_jobs):
-                    tg.start_soon(dowork, receiver.clone())
-            async with sender:
-                if dandisets:
-                    for did in dandisets:
-                        log.info("Scanning Dandiset %s", did)
-                        await sender.send(await self.get_dandiset(did))
-                else:
-                    async for ds in self.aiterdandisets():
-                        log.info("Found Dandiset %s", ds.identifier)
-                        await sender.send(ds)
-
-        if dandisets:
-            return
-
-        async with await (self.reports_root / "README.md").open("w") as fp:
-            dandiset_qtys = {
-                Outcome.PASS: 0,
-                Outcome.FAIL: 0,
-                Outcome.TIMEOUT: 0,
-            }
-            asset_qtys = {
-                Outcome.PASS: 0,
-                Outcome.FAIL: 0,
-                Outcome.TIMEOUT: 0,
-            }
-            test_summaries = {tn: TestSummary(tn) for tn in TEST_NAMES}
-            for r in all_reports:
-                passed, failed, timedout = r.combined_counts()
-                asset_qtys[Outcome.PASS] += passed
-                asset_qtys[Outcome.FAIL] += failed
-                asset_qtys[Outcome.TIMEOUT] += timedout
-                dandiset_qtys[Outcome.PASS] += bool(not failed and not timedout)
-                dandiset_qtys[Outcome.FAIL] += bool(failed)
-                dandiset_qtys[Outcome.TIMEOUT] += bool(timedout)
-                for tn in TEST_NAMES:
-                    test_summaries[tn].register(r.identifier, r.tests[tn])
-            await fp.write(
-                "| Test / (Dandisets/assets)"
-                f" | Passed ({dandiset_qtys[Outcome.PASS]}/{asset_qtys[Outcome.PASS]})"
-                f" | Failed ({dandiset_qtys[Outcome.FAIL]}/{asset_qtys[Outcome.FAIL]})"
-                f" | Timed Out ({dandiset_qtys[Outcome.TIMEOUT]}"
-                f"/{asset_qtys[Outcome.TIMEOUT]}) |\n"
-            )
-            await fp.write("| --- | --- | --- | --- |\n")
-            for tn in TEST_NAMES:
-                await fp.write(test_summaries[tn].as_row() + "\n")
-            await fp.write("\n")
-            await fp.write("| Dandiset | " + " | ".join(TEST_NAMES) + " |\n")
-            await fp.write("| --- | " + " | ".join("---" for _ in TESTS) + " |\n")
-            for did, tests in sorted((r.identifier, r.summary()) for r in all_reports):
-                await fp.write(
-                    f"| [{did}][results/{did}/status.yaml] | "
-                    + " | ".join(tests[tn] for tn in TEST_NAMES)
-                    + " |\n"
-                )
-
-    async def aiterdandisets(self) -> AsyncIterator[Dandiset]:
-        async for p in self.backup_root.iterdir():
-            if re.fullmatch(r"\d{6,}", p.name) and await p.is_dir():
-                yield await Dandiset.for_path(p)
-
-    async def get_dandiset(self, identifier: str) -> Dandiset:
-        return await Dandiset.for_path(self.backup_root / identifier)
-
-
-@dataclass
-class Dandiset:
-    identifier: str
-    path: anyio.Path
-    commit: str
-
-    @classmethod
-    async def for_path(cls, path: anyio.Path) -> Dandiset:
-        r = await anyio.run_process(["git", "show", "-s", "--format=%H"], cwd=str(path))
-        return cls(
-            identifier=path.name, path=path, commit=r.stdout.decode("utf-8").strip()
-        )
-
-    async def test_assets(self) -> DandisetReport:
-        report = DandisetReport(identifier=self.identifier, commit=self.commit)
-
-        async def dowork(rec: MemoryObjectReceiveStream[TestCase]) -> None:
-            async with rec:
-                async for testcase in rec:
-                    report.register_test_result(await testcase.run())
-
-        async with anyio.create_task_group() as tg:
-            sender, receiver = anyio.create_memory_object_stream(math.inf)
-            async with receiver:
-                for _ in range(WORKERS_PER_DANDISET):
-                    tg.start_soon(dowork, receiver.clone())
-            async with sender:
-                async for asset in self.aiterassets():
-                    log.info(
-                        "Dandiset %s: found asset %s", self.identifier, asset.asset_path
-                    )
-                    report.nassets += 1
-                    if asset.is_nwb():
-                        for t in TESTS:
-                            await sender.send(TestCase(asset=asset, testfunc=t))
-        return report
-
-    async def aiterassets(self) -> AsyncIterator[Asset]:
-        def mkasset(filepath: anyio.Path) -> Asset:
-            return Asset(
-                dandiset_id=self.identifier,
-                filepath=filepath,
-                asset_path=filepath.relative_to(self.path).as_posix(),
-            )
-
-        dirs = deque([self.path])
-        while dirs:
-            async for p in dirs.popleft().iterdir():
-                if p.name in (
-                    ".dandi",
-                    ".datalad",
-                    ".git",
-                    ".gitattributes",
-                    ".gitmodules",
-                ):
-                    continue
-                if await p.is_dir():
-                    if p.suffix in (".zarr", ".ngff"):
-                        yield mkasset(p)
-                    else:
-                        dirs.append(p)
-                else:
-                    yield mkasset(p)
-
-
-@dataclass
-class DandisetReport:
-    identifier: str
-    commit: str
-    nassets: int = 0
-    tests: dict[str, TestReport] = field(
-        default_factory=lambda: defaultdict(TestReport)
-    )
-
-    def register_test_result(self, r: TestResult) -> None:
-        self.tests[r.testname].by_outcome[r.outcome].append(r)
-
-    def combined_counts(self) -> tuple[int, int, int]:
-        asset_outcomes = defaultdict(set)
-        for report in self.tests.values():
-            for result in report.all_results():
-                asset_outcomes[result.asset.asset_path].add(result.outcome)
-        passed = 0
-        failed = 0
-        timedout = 0
-        for outcomes in asset_outcomes.values():
-            if Outcome.FAIL in outcomes:
-                failed += 1
-            if Outcome.TIMEOUT in outcomes:
-                timedout += 1
-            if outcomes == {Outcome.PASS}:
-                passed += 1
-        return (passed, failed, timedout)
-
-    def summary(self) -> dict[str, str]:
-        return {testname: self.tests[testname].summary() for testname in TEST_NAMES}
-
-    async def dump(self, reportdir: anyio.Path, run_data: RunData) -> None:
-        status = {
-            "last_run": run_data.timestamp,
-            "dandiset_version": self.commit,
-            "nassets": self.nassets,
-            "versions": run_data.versions,
-            "tests": [
-                {
-                    "name": name,
-                    "assets_ok": sorted(
-                        r.asset.asset_path for r in self.tests[name].passed
-                    ),
-                    "assets_nok": sorted(
-                        r.asset.asset_path for r in self.tests[name].failed
-                    ),
-                    "assets_timeout": sorted(
-                        r.asset.asset_path for r in self.tests[name].timedout
-                    ),
-                }
-                for name in TEST_NAMES
-                if name in self.tests
-            ],
-        }
-        yaml = YAML(typ="safe")
-        yaml.default_flow_style = False
-        out = io.StringIO()
-        yaml.dump(status, out)
-        await reportdir.mkdir(parents=True, exist_ok=True)
-        await (reportdir / "status.yaml").write_text(out.getvalue())
-        for testname, report in self.tests.items():
-            if report.failed:
-                async with await (
-                    reportdir
-                    / f"{run_data.timestamp:%Y.%m.%d.%H.%M.%S}_{testname}_errors.log"
-                ).open("w", encoding="utf-8", errors="surrogateescape") as fp:
-                    for r in report.failed:
-                        assert r.output is not None
-                        await fp.write(
-                            f"Asset: {r.asset.asset_path}\nOutput:\n"
-                            + textwrap.indent(r.output, " " * 4)
-                            + "\n"
-                        )
-
-
-@dataclass
-class TestReport:
-    by_outcome: dict[Outcome, list[TestResult]] = field(
-        init=False, default_factory=lambda: defaultdict(list)
-    )
-
-    @property
-    def passed(self) -> list[TestResult]:
-        return self.by_outcome[Outcome.PASS]
-
-    @property
-    def failed(self) -> list[TestResult]:
-        return self.by_outcome[Outcome.FAIL]
-
-    @property
-    def timedout(self) -> list[TestResult]:
-        return self.by_outcome[Outcome.TIMEOUT]
-
-    def all_results(self) -> list[TestResult]:
-        return [tr for lst in self.by_outcome.values() for tr in lst]
-
-    def counts(self) -> tuple[int, int, int]:
-        return (len(self.passed), len(self.failed), len(self.timedout))
-
-    def summary(self) -> str:
-        passed, failed, timedout = self.counts()
-        if passed == failed == timedout == 0:
-            return "\u2014"
-        else:
-            return f"{passed} passed, {failed} failed, {timedout} timed out"
-
-
-@dataclass
-class TestSummary:
-    name: str
-    dandisets_passed: int = 0
-    assets_passed: int = 0
-    dandisets_failed: dict[str, int] = field(default_factory=dict)
-    assets_failed: int = 0
-    dandisets_timedout: dict[str, int] = field(default_factory=dict)
-    assets_timedout: int = 0
-
-    def register(self, dandiset_id: str, report: TestReport) -> None:
-        passed, failed, timedout = report.counts()
-        self.assets_passed += passed
-        if failed:
-            self.dandisets_failed[dandiset_id] = failed
-            self.assets_failed += failed
-        if timedout:
-            self.dandisets_timedout[dandiset_id] = timedout
-            self.assets_timedout += timedout
-        if not failed and not timedout:
-            self.dandisets_passed += 1
-
-    def as_row(self) -> str:
-        s = f"| {self.name} | "
-        if self.dandisets_passed:
-            s += f"{self.dandisets_passed}/{self.assets_passed}"
-        else:
-            s += "\u2014"
-        s += " | "
-        if self.dandisets_failed:
-            s += f"{len(self.dandisets_failed)}/{self.assets_failed}: " + ", ".join(
-                f"[{did}]({did}/status.yaml)/{failed}"
-                for did, failed in sorted(self.dandisets_failed.items())
-            )
-        else:
-            s += "\u2014"
-        s += " | "
-        if self.dandisets_timedout:
-            s += f"{len(self.dandisets_timedout)}/{self.assets_timedout}: " + ", ".join(
-                f"[{did}]({did}/status.yaml)/{timedout}"
-                for did, timedout in sorted(self.dandisets_timedout.items())
-            )
-        else:
-            s += "\u2014"
-        s += " |"
-        return s
-
-
-@dataclass
-class Asset:
-    dandiset_id: str
-    filepath: anyio.Path
-    asset_path: str
-
-    def is_nwb(self) -> bool:
-        return self.filepath.suffix.lower() == ".nwb"
-
-
-@click.command()
+@main.command()
 @click.option(
     "-d",
     "--dataset-path",
@@ -491,7 +41,7 @@ class Asset:
     show_default=True,
 )
 @click.argument("dandisets", nargs=-1)
-def main(
+def check(
     dataset_path: anyio.Path,
     mount_point: anyio.Path,
     dandiset_jobs: int,
@@ -524,7 +74,7 @@ def main(
     matnwb_version = install_matnwb()
     hs = HealthStatus(
         backup_root=mount_point,
-        reports_root=anyio.Path(__file__).parent,
+        reports_root=anyio.Path(os.getcwd()),
     )
     hs.run_data.versions["matnwb"] = matnwb_version
     with open("fuse.log", "wb") as fp:
@@ -548,8 +98,52 @@ def main(
                 p.send_signal(SIGINT)
 
 
-def get_package_versions() -> dict[str, str]:
-    return {pkg: version(pkg) for pkg in PACKAGES_TO_VERSION}
+@main.command()
+def report() -> None:
+    r"""
+    async with await (self.reports_root / "README.md").open("w") as fp:
+        dandiset_qtys = {
+            Outcome.PASS: 0,
+            Outcome.FAIL: 0,
+            Outcome.TIMEOUT: 0,
+        }
+        asset_qtys = {
+            Outcome.PASS: 0,
+            Outcome.FAIL: 0,
+            Outcome.TIMEOUT: 0,
+        }
+        test_summaries = {tn: TestSummary(tn) for tn in TEST_NAMES}
+        for r in all_reports:
+            passed, failed, timedout = r.combined_counts()
+            asset_qtys[Outcome.PASS] += passed
+            asset_qtys[Outcome.FAIL] += failed
+            asset_qtys[Outcome.TIMEOUT] += timedout
+            dandiset_qtys[Outcome.PASS] += bool(not failed and not timedout)
+            dandiset_qtys[Outcome.FAIL] += bool(failed)
+            dandiset_qtys[Outcome.TIMEOUT] += bool(timedout)
+            for tn in TEST_NAMES:
+                test_summaries[tn].register(r.identifier, r.tests[tn])
+        await fp.write(
+            "| Test / (Dandisets/assets)"
+            f" | Passed ({dandiset_qtys[Outcome.PASS]}/{asset_qtys[Outcome.PASS]})"
+            f" | Failed ({dandiset_qtys[Outcome.FAIL]}/{asset_qtys[Outcome.FAIL]})"
+            f" | Timed Out ({dandiset_qtys[Outcome.TIMEOUT]}"
+            f"/{asset_qtys[Outcome.TIMEOUT]}) |\n"
+        )
+        await fp.write("| --- | --- | --- | --- |\n")
+        for tn in TEST_NAMES:
+            await fp.write(test_summaries[tn].as_row() + "\n")
+        await fp.write("\n")
+        await fp.write("| Dandiset | " + " | ".join(TEST_NAMES) + " |\n")
+        await fp.write("| --- | " + " | ".join("---" for _ in TESTS) + " |\n")
+        for did, tests in sorted((r.identifier, r.summary()) for r in all_reports):
+            await fp.write(
+                f"| [{did}][results/{did}/status.yaml] | "
+                + " | ".join(tests[tn] for tn in TEST_NAMES)
+                + " |\n"
+            )
+    """
+    raise NotImplementedError
 
 
 def install_matnwb() -> str:
